requirements.txt
# requirements.txt
python-telegram-bot==12.7
google-cloud-speech
google-cloud-storage
textrazor
